+++ snafu/__main__.py
@@ -24,7 +24,6 @@ def cli(ctx, version):
 @cli.command(help='Install a Python version.')
 @click.argument('version')
 @click.option('--use', is_flag=True, help='Use version after installation.')
-@click.option('--add', is_flag=True, help='Add scripts after installation.')
 @click.option(
     '--file', 'from_file', type=click.Path(exists=True),
     help='Specify an installer to not downloading one.',